requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fonttools==4.60.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain-core==0.3.76
langchain-openai==0.3.33
langsmith==0.4.29
matplotlib==3.10.6
networkx==3.5
numpy==2.3.3
openai==1.108.1
orjson==3.11.3
packaging==25.0
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
zstandard==0.25.0
